[nanvix-sandbox] E: Move Sandbox Implementations
Relocate the single_process and multi_process sandbox implementation modules
from src/libs/nanvix-sandbox/src/ into a new src/libs/nanvix-sandbox/src/sandbox/
subdirectory. This consolidation groups all process-mode-specific LinuxDaemon and
UserVm implementations under a single parent module selected at compile time via
cfg_if feature flags.

The lib.rs root module is updated to re-export from sandbox::single_process and
sandbox::multi_process instead of the previous top-level paths. A new sandbox/mod.rs
is introduced to host the feature-gated module declarations. All file renames are
pure moves with no content changes.
