+++ src/libs/nanvix-sandbox/src/sandbox/mod.rs
@@ -1,0 +1,19 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//! Sandbox implementations.
+//!
+//! This module contains the process-mode-specific sandbox implementations for Linux Daemon and
+//! User VM management, selected at compile time via feature flags.
+
+//==================================================================================================
+// Modules
+//==================================================================================================
+
+::cfg_if::cfg_if! {
+    if #[cfg(feature = "single-process")] {
+        pub mod single_process;
+    } else {
+        pub mod multi_process;
+    }
+}
